refactor: add channel type fiel and apply
