+++ client/src/component/organism/SideBar/SideBar.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import A from '@atom'
 import M from '@molecule'
 import myIcon from '@constant/icon'
-import { ChannelResponseType } from '@store/reducer/channel.reducer'
+import { ChannelResponseType } from '@type/channel.type'
 import { WorkspaceResponseType } from '@type/workspace.type'
 import Styled from './SideBar.style'
